backend/parquetformatter_api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --output-file=backend/parquetformatter_api/requirements.txt --pip-args=None backend/parquetformatter_api/requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
cachetools==6.2.0
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via uvicorn
cryptography==46.0.1
    # via pyjwt
deprecation==2.1.0
    # via
    #   postgrest
    #   storage3
duckdb==1.4.0
    # via -r backend/parquetformatter_api/requirements.in
fastapi==0.118.0
    # via -r backend/parquetformatter_api/requirements.in
google-api-core[grpc]==2.25.1
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
google-auth==2.41.0
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
google-cloud-appengine-logging==1.6.2
    # via google-cloud-logging
google-cloud-audit-log==0.3.2
    # via google-cloud-logging
google-cloud-core==2.4.3
    # via google-cloud-logging
google-cloud-logging==3.12.1
    # via -r backend/parquetformatter_api/requirements.in
googleapis-common-protos[grpc]==1.70.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-logging
grpcio==1.75.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.75.1
    # via google-api-core
gunicorn==23.0.0
    # via -r backend/parquetformatter_api/requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx[http2]==0.28.1
    # via
    #   postgrest
    #   storage3
    #   supabase
    #   supabase-auth
    #   supabase-functions
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
iniconfig==2.1.0
    # via pytest
numpy==2.3.3
    # via pandas
opentelemetry-api==1.37.0
    # via google-cloud-logging
packaging==25.0
    # via
    #   deprecation
    #   gunicorn
    #   pytest
pandas==2.3.3
    # via -r backend/parquetformatter_api/requirements.in
pluggy==1.6.0
    # via pytest
polars==1.33.1
    # via -r backend/parquetformatter_api/requirements.in
postgrest==2.20.0
    # via supabase
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-logging
protobuf==6.32.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyarrow==21.0.0
    # via -r backend/parquetformatter_api/requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.11.9
    # via
    #   fastapi
    #   postgrest
    #   realtime
    #   storage3
    #   supabase-auth
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via pytest
pyjwt[crypto]==2.10.1
    # via supabase-auth
pytest==8.4.2
    # via
    #   -r backend/parquetformatter_api/requirements.in
    #   pytest-asyncio
pytest-asyncio==1.2.0
    # via -r backend/parquetformatter_api/requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via uvicorn
python-multipart==0.0.20
    # via -r backend/parquetformatter_api/requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via uvicorn
realtime==2.20.0
    # via supabase
requests==2.32.5
    # via google-api-core
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.48.0
    # via fastapi
storage3==2.20.0
    # via supabase
strenum==0.4.15
    # via supabase-functions
supabase==2.20.0
    # via -r backend/parquetformatter_api/requirements.in
supabase-auth==2.20.0
    # via supabase
supabase-functions==2.20.0
    # via supabase
typing-extensions==4.15.0
    # via
    #   fastapi
    #   grpcio
    #   opentelemetry-api
    #   pydantic
    #   pydantic-core
    #   realtime
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.37.0
    # via -r backend/parquetformatter_api/requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via
    #   realtime
    #   uvicorn
zipp==3.23.0
    # via importlib-metadata
zipstream-ng==1.9.0
    # via -r backend/parquetformatter_api/requirements.in
